Major file refactor
- Seperate helpers.js into utils/<function-name>.js
- Seperate completion-provider private/helper methods into utils
- Rename src to lib and create top-level index.js to make library follow npm module file structure
