+++ index.js
@@ -1,0 +1,5 @@
+'use babel';
+
+const AutocompleteModulesPlugin = require('./lib/autocomplete-modules-plugin');
+
+module.exports = new AutocompleteModulesPlugin();
