+++ lib/autocomplete-modules-plugin.js
@@ -2,7 +2,7 @@
 
 const CompletionProvider = require('./completion-provider');
 
-class AutocompleteModulesPlugin {
+module.exports = class AutocompleteModulesPlugin {
   constructor() {
     this.config = {
       includeExtension: {
@@ -47,6 +47,4 @@ class AutocompleteModulesPlugin {
   getCompletionProvider() {
     return this.completionProvider;
   }
-}
-
-module.exports = new AutocompleteModulesPlugin();
+};
