+++ lib/internal-modules.js
