+++ lib/utils/lookup-webpack.js
